# perf: manually inline SortSearchWrapper usage