--- a/examples/compbench/go.mod
+++ b/examples/compbench/go.mod
@@ -1,0 +1,10 @@
+module github.com/mroth/weightedrand/examples/compbench
+
+go 1.15
+
+require (
+	github.com/jmcvetta/randutil v0.0.0-20150817122601-2bb1b664bcff
+	github.com/mroth/weightedrand v0.0.0
+)
+
+replace github.com/mroth/weightedrand => ../..