The standard library `sort.SearchInts()` is just a convenience wrapper for the generic `sort.Search()` function, which takes a function closure to determine truthfulness. However, since this function is utilized within a for loop, it cannot currently be properly inlined by the compiler, resulting in non-trivial performance overhead.

Thus, this commit introduces what is essentially a manually inlined version. In our use case here, it results in a up to ~33% overall throughput increase for `Pick()`.

```
$ benchstat before.txt after.txt
name             old time/op  new time/op  delta
Pick/10-16       44.6ns ± 7%  35.3ns ± 5%  -20.88%  (p=0.000 n=10+10)
Pick/100-16      66.6ns ± 1%  50.0ns ± 1%  -24.88%  (p=0.000 n=10+9)
Pick/1000-16     95.8ns ± 1%  64.4ns ± 1%  -32.70%  (p=0.000 n=10+10)
Pick/10000-16     122ns ± 0%    82ns ± 0%  -32.77%  (p=0.000 n=10+10)
Pick/100000-16    160ns ± 1%   105ns ± 0%  -34.50%  (p=0.000 n=10+8)
Pick/1000000-16   298ns ± 1%   219ns ± 2%  -26.50%  (p=0.000 n=10+10)
```