@mroth 2020-10-06
At this point, our single-core performance is improved enough that it may be worth redoing the benchmarks in the README.md when merging this.